Create a friendlyLoad() function.
