+++ frontend/lib/page.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import { Helmet } from "react-helmet";
-import Loadable from 'react-loadable';
 import { AriaAnnouncement } from './aria';
 
 interface PageProps {
@@ -18,17 +17,4 @@ export default function Page(props: PageProps): JSX.Element {
       {props.children}
     </React.Fragment>
   );
-}
-
-export function LoadingPage(props: Loadable.LoadingComponentProps): JSX.Element {
-  if (props.error) {
-    return (
-      <Page title="Network error">
-        <p>Unfortunately, a network error occurred.</p>
-        <br/>
-        <button className="button" onClick={props.retry}>Retry</button>
-      </Page>
-    );
-  }
-  return <Page title="Loading...">Loading...</Page>;
 }
